requirements.txt
langchain
langchain-core
langchain-community
langgraph
openai
requests
python-dotenv
PyPDF2
pypdf
scikit-learn
numpy
huggingface_hub
pydantic
